# Refactor Cesium.js to use idiomatic patterns and improve performance