## Summary
- Refactored cesium.js to use idiomatic Cesium patterns with CustomDataSource and native time-dynamic properties
- Improved terrain visualization with enhanced lighting and resolution settings
- Removed non-functional Sentinel-2 imagery provider

## Changes

### ✨ Cesium.js Refactoring
- Implemented CustomDataSource pattern for better data management
- Used native SampledPositionProperty for time-dynamic positions
- Added efficient bulk position loading
- Implemented proper performance monitoring and reporting

### 🗺️ Terrain & Visualization Improvements
- Enhanced terrain lighting with `enableLighting: true`
- Set `maximumScreenSpaceError: 1.0` for higher terrain quality
- Added high DPI display support with resolution scaling
- Improved fog settings for clearer distant terrain

### 🧹 Code Cleanup
- Removed unused Sentinel-2 imagery provider (was not loading correctly)
- Deleted obsolete files:
  - `cesium-backup.js` (1500+ lines)
  - `cesium_map.html` (old version)
  - `README_SITE_MATCHING.md`
  - Shell scripts: `flutter-android.sh`, `run.sh`
- Added cesiumLog compatibility shim for Flutter integration

### 📚 Documentation
- Added Cesium Native feasibility analysis
- Created comprehensive Product Requirements Document for Cesium visualization

## Performance Improvements
- Data processing: ~2ms for 200+ track points
- Bulk position loading: <1ms
- Static track creation: ~1ms
- Total track load time: ~13ms

## Test Plan
- [x] Test refactored Cesium.js on emulator
- [x] Test on physical device (Pixel 9)
- [x] Verify terrain rendering quality
- [x] Confirm all imagery providers work (except removed Sentinel-2)
- [x] Check performance metrics in console logs

🤖 Generated with [Claude Code](https://claude.ai/code)